Admin panel for user options(draft)

+++ src/main/java/pl/schoolmanager/repository/UserRoleRepository.java
@@ -2,8 +2,7 @@ package pl.schoolmanager.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import pl.schoolmanager.entity.User;
 import pl.schoolmanager.entity.UserRole;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Integer>{
+public interface UserRoleRepository extends JpaRepository<UserRole, Long>{
 }
